backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
python-multipart
pillow
google-generativeai
python-dotenv
